feat: add module chapter
