+++ src/modules/cjsToEsm.mjs
@@ -1,0 +1,42 @@
+import path, {dirname} from 'path';
+import { release, version } from 'os';
+import {createServer as createServerHttp} from 'http';
+import { readFile } from 'fs/promises';
+import { createRequire } from "module";
+import {fileURLToPath} from 'url';
+import './files/c.js';
+
+const random = Math.random();
+const __filename = fileURLToPath(import.meta.url);
+const __dirname = dirname(__filename);
+
+let unknownObject;
+// I find two way solve this case
+if (random > 0.5) {
+  unknownObject = JSON.parse(
+      await readFile(
+          new URL('./files/a.json', import.meta.url)
+      )
+  );
+} else {
+  const require = createRequire(import.meta.url);
+  unknownObject = require("./files/b.json");
+}
+
+console.log(unknownObject)
+console.log(`Release ${release()}`);
+console.log(`Version ${version()}`);
+console.log(`Path segment separator is "${path.sep}"`);
+
+console.log(`Path to current file is ${__filename}`);
+console.log(`Path to current directory is ${__dirname}`);
+
+const createMyServer = createServerHttp((_, res) => {
+    res.end('Request accepted');
+});
+
+export {
+    unknownObject,
+    createMyServer,
+};
+
